requirements.txt
requests[security]
PyYaml
Jinja2
